requirements.txt
gitpython
requests
tqdm
rich
